requirements.txt
aioredis==1.3.1
alabaster==0.7.8
appdirs==1.4.4
asgi-redis==1.4.3
asgiref==3.6.0
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.7.1
Automat==0.8.0
Babel==2.6.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.3
blinker==1.4
bs4==0.0.1
certifi==2019.11.28
channels==3.0.2
channels-redis==3.2.0
chardet==3.0.4
cliapp==1.20180812.1
click==8.1.2
cloud-init==22.2
cmdtest==0.32+git
colorama==0.4.6
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
daphne==3.0.1
dbus-python==1.2.16
defusedxml==0.7.1
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.2.1
django-cors-headers==3.14.0
django-environ==0.4.5
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
docopt==0.6.2
docutils==0.16
dronekit==2.9.2
dronekit-sitl==3.3.0
entrypoints==0.3
exceptiongroup==1.1.1
feedparser==5.2.1
filelock==3.0.12
Flask==2.1.1
future==0.18.2
gevent==20.12.1
gitsome==0.8.0
greenlet==0.4.17
gunicorn==20.0.4
hiredis==1.1.0
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
imagesize==1.2.0
importlib-metadata==4.11.3
incremental==16.10.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
livereload==2.6.3
lxml==4.6.2
Markdown==3.1.1
MarkupSafe==2.1.1
MAVProxy==1.8.29
monotonic==1.5
more-itertools==4.2.0
msgpack==1.0.2
msgpack-python==0.5.6
mypy==1.3.0
mypy-extensions==1.0.0
nanoid==2.0.0
netifaces==0.10.4
numpy==1.20.3
numpydoc==0.7.0
oauthlib==3.1.0
olefile==0.46
packaging==23.1
pandas==1.2.4
Pillow==7.0.0
pluggy==1.0.0
ply==3.11
prompt-toolkit==2.0.10
psutil==5.8.0
py==1.11.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pygame==2.1.2
Pygments==2.3.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
pymavlink==2.4.14
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyrsistent==0.15.5
pyserial==3.4
pytest==7.3.1
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-dotenv==0.20.0
python3-openid==3.2.0
pyttsx3==2.90
pytz==2020.4
PyYAML==5.3.1
redis==2.10.6
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
roman==2.0.0
SecretStorage==2.3.1
selenium==3.141.0
service-identity==18.1.0
setproctitle==1.1.10
simplejson==3.16.0
six==1.14.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
soupsieve==2.0.1
Sphinx==1.8.5
sqlparse==0.4.1
ssh-import-id==5.10
systemd-python==234
toml==0.10.2
tomli==2.0.1
tornado==6.1
ttystatus==0.38
Twisted==18.9.0
txaio==20.4.1
typing-extensions==4.5.0
ubuntu-advantage-tools==20.3
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
virtualenv==20.0.17
wadllib==1.3.3
wcwidth==0.1.8
websocket==0.2.1
websocket-client==0.57.0
Werkzeug==2.1.1
whitenoise==5.2.0
xonsh==0.9.13
zipp==1.0.0
zope.event==4.5.0
zope.interface==4.7.1
